renamed Arrayizers to Schema
